dashboard/requirements.txt
streamlit>=1.33
plotly
requests
pandas
python-dotenv